Add Python headers to source file
I had to do this as a workaround to ziggy-pydust passing the wrong
path to the translate c root source path, causing it trying to find
"pydust/src/ffi.h" in local projects, not Ziggy Pydust itself.

A fix should be added to "pydust/src/pydust.build.zig" in function
"addTranslateC".